Change tests: no shmin; minor fixes

+++ orchid/native_stage_adapter.py
@@ -521,9 +521,9 @@ class CreateStageDto:
         else:
             native_shmin = ScriptAdapter.MakeOptionSome(
                 onq.as_net_quantity(project_unit_system.PRESSURE, self.maybe_shmin))
-        completion_order_on_well = self.order_of_completion_on_well
+        completion_order_on_well = System.UInt32(self.order_of_completion_on_well)
         connection_type = self.connection_type.value
-        cluster_count = self.cluster_count
+        cluster_count = System.UInt32(self.cluster_count)
         no_time_range_native_stage = self.create_net_stage(well.dom_object, completion_order_on_well,
                                                            connection_type, native_md_top,
                                                            native_md_bottom, native_shmin,
